moving component processor into component pacakge
